+++ src/net/sf/mzmine/data/PeakList.java
@@ -143,7 +143,7 @@ public interface PeakList {
      * 
      * @return Maximum intensity
      */
-    public float getDataPointMaxIntensity();
+    public double getDataPointMaxIntensity();
 
     /**
      * Add a new row to the peak list
