+++ src/net/sf/mzmine/modules/dataanalysis/projectionplots/CDADataset.java
@@ -247,7 +247,7 @@ public class CDADataset extends AbstractXYDataset implements
 		return "CDA projection " + parameters.getSourcePeakList(); 
 	}	
 	
-	public float getFinishedPercentage() {
+	public double getFinishedPercentage() {
 		if (projectionStatus==null) return 0.0f;
 		return projectionStatus.getFinishedPercentage();
 	}
